@@ -41,7 +41,7 @@
   },
   "devDependencies": {
     "@cdo/apps": "file:src",
-    "@code-dot-org/blockly": "1.6.17",
+    "@code-dot-org/blockly": "1.6.18",
     "@code-dot-org/bramble": "0.1.14",
     "@code-dot-org/johnny-five": "0.10.10-cdo.0",
     "@code-dot-org/js-interpreter-tyrant": "^0.1.1",
@@ -85,7 +85,7 @@
     "ejs-loader": "0.2.1",
     "enzyme": "2.4.1",
     "es6-promise": "3.0.2",
-    "eslint": "^3.7.1",
+    "eslint": "^3.19.0",
     "eslint-plugin-mocha": "^4.9.0",
     "eslint-plugin-react": "^5.2.2",
     "exports-loader": "^0.6.3",
@@ -144,17 +144,15 @@
     "moment": "2.14.1",
     "newrelic": "^1.27.2",
     "node-sass": "^4.0.0",
-    "npm-which": "2.0.0",
     "phantomjs-prebuilt": "^2.1.14",
-    "playground-io": "code-dot-org/playground-io#v0.4.0-cdo.2",
+    "playground-io": "code-dot-org/playground-io#v0.5.0-cdo.0",
     "progress-bar-webpack-plugin": "^1.8.0",
     "pusher-js": "3.2.1",
     "radium": "0.18.1",
     "react": "15.3.2",
-    "react-addons-shallow-compare": "15.3.0",
     "react-addons-test-utils": "15.3.0",
     "react-bootstrap": "0.30.1",
-    "react-color": "2.2.5",
+    "react-color": "^2.11.7",
     "react-datepicker": "0.37.0",
     "react-dom": "15.3.0",
     "react-idle-timer": "1.3.3",
@@ -183,7 +181,6 @@
     "sinon": "^2.1.0",
     "sinon-chai": "^2.9.0",
     "sprintf-js": "^1.0.3",
-    "stackblur": "^1.0.0",
     "style-loader": "^0.13.1",
     "time-grunt": "^1.3.0",
     "time-grunt-nowatch": "github:pcardune/time-grunt-nowatch#c88a3793cf7953090f78fa36f0e991abdcd756da",
@@ -193,8 +190,7 @@
     "webpack-dev-server": "^1.14.1",
     "webpack-hot-middleware": "^2.10.0",
     "webpack-livereload-plugin": "^0.8.1",
-    "webpack-notifier": "^1.4.1",
-    "which": "1.1.1"
+    "webpack-notifier": "^1.4.1"
   },
   "dependencies": {
     "@code-dot-org/js-interpreter": "^1.1.1",
@@ -202,6 +198,7 @@
     "filesaver.js": "0.2.0",
     "jszip": "3.0.0",
     "query-string": "4.1.0",
+    "react-addons-shallow-compare": "^0.14.0 || ^15.0.0",
     "react-hot-loader": "^1.3.1",
     "selectize": "^0.12.4"
   }
