@@ -55,29 +55,35 @@ const styles = {
 
 const GradientNavCard = React.createClass({
   propTypes: {
-    cardData: React.PropTypes.shape({
-      title: React.PropTypes.string.isRequired,
-      description: React.PropTypes.string.isRequired,
-      image: React.PropTypes.string.isRequired,
-      buttonText: React.PropTypes.string.isRequired,
-      link: React.PropTypes.string.isRequired
-    })
+    title: React.PropTypes.string.isRequired,
+    description: React.PropTypes.string.isRequired,
+    image: React.PropTypes.string.isRequired,
+    buttonText: React.PropTypes.string.isRequired,
+    link: React.PropTypes.string.isRequired
   },
 
   render() {
-    const { cardData } = this.props;
+    const { title, description, buttonText, link } = this.props;
 
     return (
       <div style={styles.card}>
         <img src={require('../../../static/navcard-placeholder.png')} style={styles.image}/>
-        <h2 style={styles.title}>{cardData.title}</h2>
-        <h4 style={styles.description}>{cardData.description}</h4>
-        <ProgressButton href={cardData.link} color="gray" text={cardData.buttonText} style={styles.button}/>
+        <div style={styles.title}>
+          {title}
+        </div>
+        <div style={styles.description}>
+          {description}
+        </div>
+        <ProgressButton
+          href={link}
+          color={ProgressButton.ButtonColor.gray}
+          text={buttonText}
+          style={styles.button}
+        />
         <div style={styles.overlay}/>
       </div>
     );
   }
-
 });
 
 export default GradientNavCard;
