@@ -1,6 +1,19 @@
-// This config only defines globals available especially
-// in apps.  See the root .eslintrc.js for linting rules.
+// This config only defines globals available especially in apps,
+// and enables es6. See the root .eslintrc.js for linting rules.
 module.exports = {
+  "parser": "babel-eslint",
+  "parserOptions": {
+    "sourceType": "module",
+    "ecmaFeatures": {
+      "jsx": true,
+      "modules": true,
+      "ecmaVersion": 6,
+      "experimentalObjectRestSpread": true
+    }
+  },
+  "env": {
+    "es6": true
+  },
   "globals": {
     "Blockly": true,
     "Phaser": true,
