@@ -449,8 +449,9 @@ testsContext.keys().forEach(testsContext);
     // tutorialExplorer for code.org/learn 2016 edition.
     tutorialExplorer: './src/tutorialExplorer/tutorialExplorer.js',
 
-    // Hamburger for header on dashboard & pegasus.
-    hamburger: './src/hamburger/hamburger.js',
+    // common entry points for pegasus js
+    'code.org/views/theme_common_head_after': './src/sites/code.org/pages/views/theme_common_head_after.js',
+    'hourofcode.com/views/theme_common_head_after': './src/sites/hourofcode.com/pages/views/theme_common_head_after.js',
 
     pd: './src/code-studio/pd/workshop_dashboard/workshop_dashboard.jsx',
 
@@ -460,8 +461,8 @@ testsContext.keys().forEach(testsContext);
 
     'pd/professional_learning_landing/index': './src/sites/studio/pages/pd/professional_learning_landing/index.js',
 
-    'teacher-dashboard/index': './src/sites/code.org/pages/teacher-dashboard/index.js',
-    'pd-workshop-survey/splat': './src/sites/code.org/pages/pd-workshop-survey/splat.js',
+    'code.org/public/teacher-dashboard/index': './src/sites/code.org/pages/public/teacher-dashboard/index.js',
+    'code.org/public/pd-workshop-survey/splat': './src/sites/code.org/pages/public/pd-workshop-survey/splat.js',
 
     publicKeyCryptography: './src/publicKeyCryptography/main.js',
 
@@ -803,17 +804,6 @@ testsContext.keys().forEach(testsContext);
     'concat',
     'karma:all'
   ]);
-
-  // We used to use 'mochaTest' as our test command.  Alias to be friendly while
-  // we transition away from it.  This can probably be removed in a month or two.
-  // - Brad (16 May 2016)
-  grunt.registerTask('showMochaTestWarning', function () {
-    console.log(chalk.yellow('Warning: ') + 'The ' + chalk.italic('mochaTest') +
-        ' task is deprecated.  Use ' + chalk.italic('test') + ' instead, or' +
-        ' directly invoke its subtasks ' + chalk.italic('unitTest') + ' and ' +
-        chalk.italic('integrationTest') + '.');
-  });
-  grunt.registerTask('mochaTest', ['showMochaTestWarning', 'test']);
 
   grunt.registerTask('logBuildTimes', function () {
     var done = this.async();
