@@ -94,9 +94,9 @@ Blockly.Msg.FUNCTIONAL_NAME_LABEL = "Ad";
 Blockly.Msg.FUNCTIONAL_TYPE_LABEL = "Tipini seçin...";
 Blockly.Msg.FUNCTIONAL_DESCRIPTION_LABEL = "Təsvir";
 Blockly.Msg.FUNCTIONAL_DOMAIN_LABEL = "Təyin oblastı";
-Blockly.Msg.FUNCTIONAL_PROCEDURE_DEFINE_TOOLTIP = "Funksional bir metod təyin et";
+Blockly.Msg.FUNCTIONAL_PROCEDURE_DEFINE_TOOLTIP = "Funksional metod təyin et";
 Blockly.Msg.FUNCTIONAL_RANGE_LABEL = "Qiymətlər oblastı";
-Blockly.Msg.FUNCTIONAL_VARIABLE_CREATE = "Bir dəyişən düzəlt";
+Blockly.Msg.FUNCTIONAL_VARIABLE_CREATE = "Bir dəyişən yarat";
 Blockly.Msg.FUNCTIONAL_VARIABLE_HEADER = "Dəyişən";
 Blockly.Msg.FUNCTIONAL_VARIABLE_TYPE = "Tip";
 Blockly.Msg.HELP = "Kömək";
@@ -177,11 +177,11 @@ Blockly.Msg.LOGIC_BOOLEAN_HELPURL = "http://code.google.com/p/blockly/wiki/True_
 Blockly.Msg.LOGIC_BOOLEAN_TOOLTIP = "\"Doğru\" və ya \"yalan\" cavabını qaytarır.";
 Blockly.Msg.LOGIC_BOOLEAN_TRUE = "doğru";
 Blockly.Msg.LOGIC_COMPARE_HELPURL = "http://en.wikipedia.org/wiki/Inequality_(mathematics)";
-Blockly.Msg.LOGIC_COMPARE_TOOLTIP_EQ = "Girişlər bir birinə bərabərdirsə \"doğru\" cavabını qaytarır.";
+Blockly.Msg.LOGIC_COMPARE_TOOLTIP_EQ = "Girişlər bir birinə bərabərdirsə, \"doğru\" cavabını qaytarır.";
 Blockly.Msg.LOGIC_COMPARE_TOOLTIP_GT = "Birinci giriş ikincidən böyükdürsə \"doğru\" cavabını qaytarır.";
-Blockly.Msg.LOGIC_COMPARE_TOOLTIP_GTE = "Birinci giriş ikincidən böyük və ya bərarbərdirsə \"doğru\" cavabını qaytarır.";
+Blockly.Msg.LOGIC_COMPARE_TOOLTIP_GTE = "Birinci giriş ikincidən böyük və ya bərarbərdirsə, \"doğru\" cavabını qaytarır.";
 Blockly.Msg.LOGIC_COMPARE_TOOLTIP_LT = "Birinci giriş ikincidən kiçikdirsə \"doğru\" cavabını qaytarır.";
-Blockly.Msg.LOGIC_COMPARE_TOOLTIP_LTE = "Birinci giriş ikincidən kiçik və ya bərarbərdirsə \"doğru\" cavabını qaytarır.";
+Blockly.Msg.LOGIC_COMPARE_TOOLTIP_LTE = "Birinci giriş ikincidən kiçik və ya bərarbərdirsə, \"doğru\" cavabını qaytarır.";
 Blockly.Msg.LOGIC_COMPARE_TOOLTIP_NEQ = "Daxil olunan dəyərlər bir-birlərinə bərabər deyilsə, doğru cavabını ötür.";
 Blockly.Msg.LOGIC_NEGATE_HELPURL = "http://code.google.com/p/blockly/wiki/Not";
 Blockly.Msg.LOGIC_NEGATE_TITLE = "%1 deyil";
@@ -192,8 +192,8 @@ Blockly.Msg.LOGIC_NULL_TOOLTIP = "Boş cavab qaytarır.";
 Blockly.Msg.LOGIC_OPERATION_AND = "və";
 Blockly.Msg.LOGIC_OPERATION_HELPURL = "http://code.google.com/p/blockly/wiki/And_Or";
 Blockly.Msg.LOGIC_OPERATION_OR = "və ya";
-Blockly.Msg.LOGIC_OPERATION_TOOLTIP_AND = "Hər iki giriş \"doğru\"-dursa \"doğru\" cavabını qaytarır.";
-Blockly.Msg.LOGIC_OPERATION_TOOLTIP_OR = "Girişlərdən heç olmasa biri \"doğru\"-dursa \"doğru\" cavabını qaytarır.";
+Blockly.Msg.LOGIC_OPERATION_TOOLTIP_AND = "Hər iki giriş \"doğru\"-dursa, \"doğru\" cavabını qaytarır.";
+Blockly.Msg.LOGIC_OPERATION_TOOLTIP_OR = "Girişlərdən heç olmasa biri \"doğru\"-dursa, \"doğru\" cavabını qaytarır.";
 Blockly.Msg.LOGIC_TERNARY_CONDITION = "yoxla";
 Blockly.Msg.LOGIC_TERNARY_HELPURL = "http://en.wikipedia.org/wiki/%3F:";
 Blockly.Msg.LOGIC_TERNARY_IF_FALSE = "əgər yalandırsa";
@@ -306,15 +306,15 @@ Blockly.Msg.PROCEDURES_DEFRETURN_RETURN = "qaytar";
 Blockly.Msg.PROCEDURES_DEFRETURN_TOOLTIP = "Nəticəsi olan funksiya yaradır.";
 Blockly.Msg.PROCEDURES_DEF_DUPLICATE_WARNING = "Xəbərdarlıq: Bu funksiyanın təkrar olunmuş parametrləri var.";
 Blockly.Msg.PROCEDURES_HIGHLIGHT_DEF = "Funksiyanın təyinatını vurğula";
-Blockly.Msg.PROCEDURES_IFRETURN_TOOLTIP = "Əgər bir dəyər \"doğru\"-dursa onda ikinci dəyəri qaytar.";
-Blockly.Msg.PROCEDURES_IFRETURN_WARNING = "Xəbərdarlıq: Bu blok ancaq bir funksiyanın təyinatı daxilində işlədilə bilər.";
+Blockly.Msg.PROCEDURES_IFRETURN_TOOLTIP = "Əgər qiymət \"doğru\"-dursa onda ikinci dəyəri qaytar.";
+Blockly.Msg.PROCEDURES_IFRETURN_WARNING = "Xəbərdarlıq: Bu blok ancaq bir funksiyanın təyinatı daxilində istifadə oluna bilər.";
 Blockly.Msg.PROCEDURES_MUTATORARG_TITLE = "Giriş adı:";
 Blockly.Msg.PROCEDURES_MUTATORCONTAINER_TITLE = "girişlər";
 Blockly.Msg.REMOVE_COMMENT = "Şərhi sil";
 Blockly.Msg.RENAME_PARAMETER = "Parametrin adını dəyiş...";
-Blockly.Msg.RENAME_PARAMETER_TITLE = "Bütün '%1' parametrlərinin adını dəyiş:";
-Blockly.Msg.RENAME_VARIABLE = "Dəyişənin adını dəyiş...";
-Blockly.Msg.RENAME_VARIABLE_TITLE = "Bütün '%1' dəyişənlərinin adını buna dəyiş:";
+Blockly.Msg.RENAME_PARAMETER_TITLE = "Bütün '%1' parametrlərinin adını dəyişdir:";
+Blockly.Msg.RENAME_VARIABLE = "Dəyişənin adını dəyişdir...";
+Blockly.Msg.RENAME_VARIABLE_TITLE = "Bütün '%1' dəyişənlərinin adını buna dəyişdir:";
 Blockly.Msg.TEXT_APPEND_APPENDTEXT = "bu mətni əlavə et:";
 Blockly.Msg.TEXT_APPEND_HELPURL = "https://code.google.com/p/blockly/wiki/Text#Text_modification";
 Blockly.Msg.TEXT_APPEND_TO = "bu mətnin sonuna:";
