@@ -200,7 +200,7 @@ Blockly.Msg.LOGIC_TERNARY_IF_FALSE = "si es falso";
 Blockly.Msg.LOGIC_TERNARY_IF_TRUE = "si es verdadero";
 Blockly.Msg.LOGIC_TERNARY_TOOLTIP = "Comprobar la condición en \"probar\". Si la condición es verdadera, devolver el valor \"si es cierto\"; de lo contrario, devolver el valor \"si es falso\".";
 Blockly.Msg.MATH_ADDITION_SYMBOL = "+";
-Blockly.Msg.MATH_ARITHMETIC_HELPURL = "http://es.wikipedia.org/wiki/Aritmética";
+Blockly.Msg.MATH_ARITHMETIC_HELPURL = "http://en.wikipedia.org/wiki/Arithmetic";
 Blockly.Msg.MATH_ARITHMETIC_TOOLTIP_ADD = "Devuelve la suma de dos números.";
 Blockly.Msg.MATH_ARITHMETIC_TOOLTIP_DIVIDE = "Devuelve el cociente de dos números.";
 Blockly.Msg.MATH_ARITHMETIC_TOOLTIP_MINUS = "Devuelve la diferencia de dos números.";
