@@ -312,9 +312,9 @@ Blockly.Msg.PROCEDURES_MUTATORARG_TITLE = "girdi adı:";
 Blockly.Msg.PROCEDURES_MUTATORCONTAINER_TITLE = "girdiler";
 Blockly.Msg.REMOVE_COMMENT = "Yorumu sil";
 Blockly.Msg.RENAME_PARAMETER = "Parametreyi yeniden adlandır...";
-Blockly.Msg.RENAME_PARAMETER_TITLE = "Tüm '%1' parametrelerini şu şekilde adlandır:";
+Blockly.Msg.RENAME_PARAMETER_TITLE = "Tüm '%1' parametreleri yeniden adlandırın:";
 Blockly.Msg.RENAME_VARIABLE = "Değişkeni yeniden adlandır...";
-Blockly.Msg.RENAME_VARIABLE_TITLE = "Tüm '%1' değişkenlerini şu şekilde adlandır:";
+Blockly.Msg.RENAME_VARIABLE_TITLE = "Tüm '%1' değişkenlerini yeniden isimlendir:";
 Blockly.Msg.TEXT_APPEND_APPENDTEXT = "metin ekle";
 Blockly.Msg.TEXT_APPEND_HELPURL = "https://code.google.com/p/blockly/wiki/Text#Text_modification   Yazı değişimi";
 Blockly.Msg.TEXT_APPEND_TO = "e";
