@@ -81,18 +81,25 @@ const SILENCED = [
   'studio',
   'turtle',
   'tutorialExplorer',
-  'hamburger',
   'weblab',
-  'learn/index'
+  'learn/index',
 ];
 const SITES_CONFIG = {
   'studio': {
+    entryPrefix: '',
     templateRoot: '../dashboard/app/views',
     templateGlobs: ['**/*.erb', '**/*.haml'],
     templateExtensions: ['erb', 'haml'],
   },
   'code.org': {
-    templateRoot: '../pegasus/sites.v3/code.org/public',
+    entryPrefix: 'code.org/',
+    templateRoot: '../pegasus/sites.v3/code.org',
+    templateGlobs: ['**/*.erb', '**/*.haml'],
+    templateExtensions: ['erb', 'haml'],
+  },
+  'hourofcode.com': {
+    entryPrefix: 'hourofcode.com/',
+    templateRoot: '../pegasus/sites.v3/hourofcode.com',
     templateGlobs: ['**/*.erb', '**/*.haml'],
     templateExtensions: ['erb', 'haml'],
   },
@@ -160,7 +167,7 @@ function checkEntryPoint(entryKey, entryPointPath, stats, options) {
         const matchedTemplatePaths = [];
         templates.forEach(templatePath => {
           const relativePath = templatePath.replace(siteConfig.templateRoot, '').slice(1);
-          possibleValidEntryKeys.add(relativePath.split('.')[0]);
+          possibleValidEntryKeys.add(siteConfig.entryPrefix + relativePath.split('.')[0]);
           matchedTemplatePaths.push(relativePath);
         });
 
@@ -184,8 +191,14 @@ function checkEntryPoint(entryKey, entryPointPath, stats, options) {
         if (entryPointPatternMatch) {
           // entry point is in the sites/<site-name>/pages direcotory (good)
           // but it doesn't have the same name as the js file it points to (bad)
-          if (entryPointPatternMatch[2] !== entryKey) {
-            errors.push(`Entry points should have the same name as the file they point to!`);
+          if (siteConfig.entryPrefix + entryPointPatternMatch[2] !== entryKey) {
+            errors.push(
+              `Entry points should have the same name as the file they point to!\n` +
+              `This entry point should be renamed to ` +
+              chalk.underline(siteConfig.entryPrefix + entryPointPatternMatch[2]) +
+              `!`
+            );
+            errors.push();
           }
         } else {
           errors.push(
