@@ -1,8 +1,11 @@
+/* eslint-disable react/no-danger */
 import React, { PropTypes } from 'react';
 import ProgressLevelSet from './ProgressLevelSet';
 import ProgressBubbleSet from './ProgressBubbleSet';
 import { levelType } from './progressTypes';
 import { progressionsFromLevels } from '@cdo/apps/code-studio/progressRedux';
+import marked from 'marked';
+import renderer from '@cdo/apps/util/StylelessRenderer';
 
 const styles = {
   summary: {
@@ -45,11 +48,16 @@ const ProgressLessonContent = React.createClass({
       );
     }
 
+    // We use dangerouslySetInnerHTML, but depend on our server markdown renderer
+    // not providing for support for most markdown features (backticks being the
+    // exception). As such, we should not end up with actually dangerous markdown
+    // here.
     return (
       <div>
-        <div style={styles.summary}>
-          {description}
-        </div>
+        <div
+          style={styles.summary}
+          dangerouslySetInnerHTML={{ __html: marked(description || '', { renderer }) }}
+        />
         {bubbles}
       </div>
     );
