@@ -6,7 +6,7 @@ module TestRunUtils
   def self.run_apps_tests
     Dir.chdir(apps_dir) do
       ChatClient.wrap('apps tests') do
-        RakeUtils.system_stream_output 'npm run test-low-memory'
+        RakeUtils.system 'npm run test-low-memory'
       end
     end
   end
