@@ -171,6 +171,135 @@ class SectionApiHelperTest < SequelTestCase
       FakeDashboard.use_fake_database
     end
 
+    describe 'teachers' do
+      it 'returns an array of hashes of information' do
+        pegasus_section = DashboardSection.fetch_if_teacher(
+          FakeDashboard::SECTION_NORMAL[:id],
+          FakeDashboard::TEACHER[:id]
+        )
+        teachers = pegasus_section.teachers
+        assert_equal(
+          [{id: FakeDashboard::TEACHER[:id], location: "/v2/users/#{FakeDashboard::TEACHER[:id]}"}],
+          teachers
+        )
+      end
+    end
+
+    describe 'teacher?' do
+      it 'returns false for a teacher of a different section' do
+        pegasus_section = DashboardSection.fetch_if_teacher(
+          FakeDashboard::SECTION_NORMAL[:id],
+          FakeDashboard::TEACHER[:id]
+        )
+        refute pegasus_section.teacher?(FakeDashboard::TEACHER_SELF[:id])
+      end
+
+      it 'returns true for a teacher of the section' do
+        pegasus_section = DashboardSection.fetch_if_teacher(
+          FakeDashboard::SECTION_NORMAL[:id],
+          FakeDashboard::TEACHER[:id]
+        )
+        assert pegasus_section.teacher?(FakeDashboard::TEACHER[:id])
+      end
+
+      it 'returns false for soft-deleted enrollments' do
+        pegasus_section = DashboardSection.fetch_if_teacher(
+          FakeDashboard::SECTION_DELETED_FOLLOWER[:id],
+          FakeDashboard::TEACHER_DELETED_FOLLOWER[:id]
+        )
+        refute pegasus_section.student?(FakeDashboard::STUDENT_DELETED_FOLLOWER[:id])
+      end
+    end
+
+    describe 'students' do
+      it 'returns hash for section with student without progress' do
+        Dashboard.db.transaction(rollback: :always) do
+          pegasus_section = DashboardSection.fetch_if_teacher(
+            FakeDashboard::SECTION_NORMAL[:id],
+            FakeDashboard::TEACHER[:id]
+          )
+          students = pegasus_section.students
+          assert_equal(
+            [
+              {
+                id: FakeDashboard::STUDENT[:id],
+                name: FakeDashboard::STUDENT[:name],
+                username: nil,
+                email: '',
+                hashed_email: nil,
+                user_type: 'student',
+                gender: nil,
+                birthday: nil,
+                prize_earned: false,
+                total_lines: 0,
+                secret_words: nil,
+                secret_picture_name: nil,
+                secret_picture_path: nil,
+                location: "/v2/users/#{FakeDashboard::STUDENT[:id]}",
+                age: nil,
+                completed_levels_count: 0
+              }
+            ],
+            students
+          )
+        end
+      end
+
+      it 'returns hash for section with student with progress' do
+        Dashboard.db.transaction(rollback: :always) do
+          Dashboard.db[:user_levels].insert(
+            user_id: FakeDashboard::STUDENT[:id],
+            level_id: 1,
+            best_result: 100
+          )
+          pegasus_section = DashboardSection.fetch_if_teacher(
+            FakeDashboard::SECTION_NORMAL[:id],
+            FakeDashboard::TEACHER[:id]
+          )
+          students = pegasus_section.students
+          assert_equal(
+            [
+              {
+                id: FakeDashboard::STUDENT[:id],
+                name: FakeDashboard::STUDENT[:name],
+                username: nil,
+                email: '',
+                hashed_email: nil,
+                user_type: 'student',
+                gender: nil,
+                birthday: nil,
+                prize_earned: false,
+                total_lines: 0,
+                secret_words: nil,
+                secret_picture_name: nil,
+                secret_picture_path: nil,
+                location: "/v2/users/#{FakeDashboard::STUDENT[:id]}",
+                age: nil,
+                completed_levels_count: 1
+              }
+            ],
+            students
+          )
+        end
+      end
+
+      it 'ignores deleted followers' do
+        pegasus_section = DashboardSection.fetch_if_teacher(
+          FakeDashboard::SECTION_DELETED_FOLLOWER[:id],
+          FakeDashboard::TEACHER_DELETED_FOLLOWER[:id]
+        )
+        assert_equal [], pegasus_section.students
+      end
+
+      it 'returns empty array for empty section' do
+        pegasus_section = DashboardSection.fetch_if_teacher(
+          FakeDashboard::SECTION_EMPTY[:id],
+          FakeDashboard::TEACHER[:id]
+        )
+        assert_equal [], pegasus_section.students
+      end
+    end
+
     describe 'remove_student' do
       it 'soft-deletes follower' do
         Dashboard.db.transaction(rollback: :always) do
@@ -237,6 +366,42 @@ class SectionApiHelperTest < SequelTestCase
     before do
       FakeDashboard.use_fake_database
       @script_id = 1
+      @time_in_past = Time.new(2017, 1, 2)
+    end
+
+    describe 'assign_script_to_section' do
+      it 'noops for an empty section' do
+        Dashboard.db.transaction(rollback: :always) do
+          DashboardUserScript.assign_script_to_section(
+            @script_id,
+            FakeDashboard::SECTION_EMPTY[:id]
+          )
+          assert_equal 0, Dashboard.db[:user_scripts].count
+        end
+      end
+
+      it 'assigns scripts to followers' do
+        Dashboard.db.transaction(rollback: :always) do
+          DashboardUserScript.assign_script_to_section(
+            @script_id,
+            FakeDashboard::SECTION_NORMAL[:id]
+          )
+          user_scripts = Dashboard.db[:user_scripts].all
+          assert_equal 1, user_scripts.count
+          assert_equal FakeDashboard::STUDENT[:id], user_scripts.first[:user_id]
+        end
+      end
+
+      it 'ignores deleted followers' do
+        Dashboard.db.transaction(rollback: :always) do
+          DashboardUserScript.assign_script_to_section(
+            @script_id,
+            FakeDashboard::SECTION_DELETED_FOLLOWER[:id]
+          )
+          user_scripts = Dashboard.db[:user_scripts].all
+          assert_equal 0, user_scripts.count
+        end
+      end
     end
 
     describe 'assign_script_to_user' do
@@ -250,16 +415,19 @@ class SectionApiHelperTest < SequelTestCase
             where(user_id: user_id, script_id: @script_id).
             all
           assert_equal 1, user_scripts.count
-          # TODO(asher): Uncomment this assertion after fixing the method to set assigned_at.
-          # assert user_scripts.first[:assigned_at]
+          assert user_scripts.first[:assigned_at]
         end
       end
 
-      it 'does not update user_scripts if one exists' do
+      it 'assigns user_scripts if one exists without assigned_at' do
         Dashboard.db.transaction(rollback: :always) do
           user_id = FakeDashboard::STUDENT[:id]
-          Dashboard.db[:user_scripts].
-            insert(user_id: user_id, script_id: @script_id, created_at: '2017-01-02 00:00:00')
+          Dashboard.db[:user_scripts].insert(
+            user_id: user_id,
+            script_id: @script_id,
+            created_at: @time_in_past,
+            updated_at: @time_in_past
+          )
 
           DashboardUserScript.assign_script_to_user(@script_id, user_id)
 
@@ -267,8 +435,31 @@ class SectionApiHelperTest < SequelTestCase
             where(user_id: user_id, script_id: @script_id).
             all
           assert_equal 1, user_scripts.count
-          # TODO(asher): Uncomment this assertion after fixing the method to set assigned_at.
-          # assert user_scripts.first[:assigned_at]
+          assert_equal @time_in_past, user_scripts.first[:created_at]
+          assert user_scripts.first[:updated_at] > @time_in_past
+          assert user_scripts.first[:assigned_at]
+        end
+      end
+
+      it 'does not update user_scripts if one exists with assigned_at' do
+        Dashboard.db.transaction(rollback: :always) do
+          user_id = FakeDashboard::STUDENT[:id]
+          Dashboard.db[:user_scripts].insert(
+            user_id: user_id,
+            script_id: @script_id,
+            created_at: @time_in_past,
+            updated_at: @time_in_past,
+            assigned_at: @time_in_past
+          )
+
+          DashboardUserScript.assign_script_to_user(@script_id, user_id)
+
+          user_scripts = Dashboard.db[:user_scripts].
+            where(user_id: user_id, script_id: @script_id).
+            all
+          assert_equal 1, user_scripts.count
+          assert_equal @time_in_past, user_scripts.first[:updated_at]
+          assert_equal @time_in_past, user_scripts.first[:assigned_at]
         end
       end
     end
@@ -295,19 +486,22 @@ class SectionApiHelperTest < SequelTestCase
             where(user_id: user_ids, script_id: @script_id).
             all
           assert_equal 2, user_scripts.count
-          # TODO(asher): Uncomment this assertion after fixing the method to set assigned_at.
-          # user_scripts.each do |user_script|
-          #   assert user_script[:assigned_at]
-          # end
+          user_scripts.each do |user_script|
+            assert user_script[:assigned_at]
+          end
         end
       end
 
-      it 'does not update user_scripts if already exists' do
+      it 'assigns user_scripts if exists without assigned_at' do
         Dashboard.db.transaction(rollback: :always) do
           user_ids = [FakeDashboard::STUDENT[:id], FakeDashboard::STUDENT_SELF[:id]]
           user_ids.each do |user_id|
-            Dashboard.db[:user_scripts].
-              insert(user_id: user_id, script_id: @script_id, created_at: '2017-01-02 00:00:00')
+            Dashboard.db[:user_scripts].insert(
+              user_id: user_id,
+              script_id: @script_id,
+              created_at: @time_in_past,
+              updated_at: @time_in_past
+            )
           end
 
           DashboardUserScript.assign_script_to_users(@script_id, user_ids)
@@ -316,10 +510,36 @@ class SectionApiHelperTest < SequelTestCase
             where(user_id: user_ids, script_id: @script_id).
             all
           assert_equal 2, user_scripts.count
-          # TODO(asher): Uncomment this assertion after fixing the method to set assigned_at.
-          # user_scripts.each do |user_script|
-          #   assert user_script[:assigned_at]
-          # end
+          user_scripts.each do |user_script|
+            assert user_script[:assigned_at]
+            assert user_script[:updated_at] > @time_in_past
+          end
+        end
+      end
+
+      it 'does not update user_scripts if exists with assigned_at' do
+        Dashboard.db.transaction(rollback: :always) do
+          user_ids = [FakeDashboard::STUDENT[:id], FakeDashboard::STUDENT_SELF[:id]]
+          user_ids.each do |user_id|
+            Dashboard.db[:user_scripts].insert(
+              user_id: user_id,
+              script_id: @script_id,
+              created_at: @time_in_past,
+              updated_at: @time_in_past,
+              assigned_at: @time_in_past
+            )
+          end
+
+          DashboardUserScript.assign_script_to_users(@script_id, user_ids)
+
+          user_scripts = Dashboard.db[:user_scripts].
+            where(user_id: user_ids, script_id: @script_id).
+            all
+          assert_equal 2, user_scripts.count
+          user_scripts.each do |user_script|
+            assert_equal @time_in_past, user_script[:assigned_at]
+            assert_equal @time_in_past, user_script[:updated_at]
+          end
         end
       end
     end
